+++ src/frontend/apps/impress/src/pages/docs/[id]/index.tsx
@@ -1,3 +1,4 @@
+import { TreeProvider } from '@gouvfr-lasuite/ui-kit';
 import { Loader } from '@openfun/cunningham-react';
 import { useQueryClient } from '@tanstack/react-query';
 import Head from 'next/head';
@@ -7,14 +8,21 @@ import { useTranslation } from 'react-i18next';
 
 import { Box, Text, TextErrors } from '@/components';
 import { DocEditor } from '@/docs/doc-editor';
+import { KEY_AUTH, setAuthUrl } from '@/features/auth';
 import {
   Doc,
   KEY_DOC,
+  KEY_SUB_PAGE,
+  getDoc,
   useCollaboration,
   useDoc,
   useDocStore,
-} from '@/docs/doc-management/';
-import { KEY_AUTH, setAuthUrl } from '@/features/auth';
+} from '@/features/docs/doc-management/';
+import {
+  getDocChildren,
+  serializeDocToSubPage,
+  subPageToTree,
+} from '@/features/docs/doc-tree/';
 import { MainLayout } from '@/layouts';
 import { useBroadcastStore } from '@/stores';
 import { NextPageWithLayout } from '@/types/next';
@@ -23,6 +31,7 @@ export function DocLayout() {
   const {
     query: { id },
   } = useRouter();
+  const queryClient = useQueryClient();
 
   if (typeof id !== 'string') {
     return null;
@@ -34,9 +43,23 @@ export function DocLayout() {
         <meta name="robots" content="noindex" />
       </Head>
 
-      <MainLayout>
-        <DocPage id={id} />
-      </MainLayout>
+      <TreeProvider
+        initialNodeId={id}
+        onRefresh={async (docId: string) => {
+          const doc = await getDoc({ id: docId });
+          void queryClient.setQueryData([KEY_DOC, { id: docId }], doc);
+          void queryClient.setQueryData([KEY_SUB_PAGE, { id: docId }], doc);
+          return serializeDocToSubPage(doc);
+        }}
+        onLoadChildren={async (docId: string) => {
+          const doc = await getDocChildren({ docId });
+          return subPageToTree(doc.results);
+        }}
+      >
+        <MainLayout>
+          <DocPage id={id} />
+        </MainLayout>
+      </TreeProvider>
     </>
   );
 }
@@ -83,6 +106,12 @@ const DocPage = ({ id }: DocProps) => {
     setDoc(docQuery);
     setCurrentDoc(docQuery);
   }, [docQuery, setCurrentDoc, isFetching]);
+
+  useEffect(() => {
+    return () => {
+      setCurrentDoc(undefined);
+    };
+  }, [setCurrentDoc]);
 
   /**
    * We add a broadcast task to reset the query cache