+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+## Added
+
+- ✨(frontend) add custom css style #771
+- ✨(frontend) multi-pages #701
+
 ## Changed
 
 - 🚨(frontend) block button when creating doc #749
@@ -21,7 +26,6 @@ and this project adheres to
 ## Added
 
 - 📄(legal) Require contributors to sign a DCO #779
-- ✨(frontend) add custom css style #771
 
 ## Changed